requirements.txt
streamlit==1.38.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain core ecosystem
langchain-core==0.2.43
langchain-community==0.2.17
langgraph==0.2.20
langchain-ollama==0.1.0

# Other libraries
spacy==3.7.2
vaderSentiment==3.3.2
pydantic==2.8.2
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.1
